% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{compare_coef}
\alias{compare_coef}
\title{Compare estimated to simulated parameters}
\usage{
compare_coef(sim, fit, coef.name, plot = TRUE)
}
\arguments{
\item{sim}{A \code{splatter} pop sim}

\item{fit}{An \code{edgeR} fit object}

\item{coef.name}{Name of the parameter to compare}

\item{plot}{Logical; make a plot?}
}
\description{
Compare estimated to simulated parameters
}
