% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{create_pb_dispersion}
\alias{create_pb_dispersion}
\title{Create pseudobulk dispersion matrix}
\usage{
create_pb_dispersion(y)
}
\arguments{
\item{y}{A \code{DGEList} object of single cell counts with fitted dispersions}
}
\description{
Create pseudobulk dispersion matrix
}
