% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{create_pseudobulk_samples}
\alias{create_pseudobulk_samples}
\title{Aggregate single cell counts to create pseudobulk samples}
\usage{
create_pseudobulk_samples(y)
}
\arguments{
\item{y}{A \code{DGEList} object}
}
\description{
Aggregate single cell counts to create pseudobulk samples
}
