% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sc_pseudodata.R
\name{generate_sc_pseudodata.DGEList}
\alias{generate_sc_pseudodata.DGEList}
\title{Generate pseudodata from single cell counts stored in a \code{DGEList}}
\usage{
\method{generate_sc_pseudodata}{DGEList}(
  y,
  prior.df = 10,
  niter = 4,
  tol = 1e-06,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{A \code{DGEList} object}
}
\description{
Generate pseudodata from single cell counts stored in a \code{DGEList}
}
