% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sc_pseudodata.R
\name{generate_sc_pseudodata}
\alias{generate_sc_pseudodata}
\title{Generate pseudodata from single cell counts}
\usage{
generate_sc_pseudodata(y, ...)
}
\arguments{
\item{y}{A matrix or \code{DGEList} object}
}
\description{
Generate pseudodata from single cell counts
}
