% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{get_coefs.DGEGLM}
\alias{get_coefs.DGEGLM}
\title{Get estimated coefficient values for cluster and condition}
\usage{
\method{get_coefs}{DGEGLM}(y)
}
\arguments{
\item{y}{A \code{DGEGLM} object}
}
\description{
Get estimated coefficient values for cluster and condition
}
