% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{get_coefs.SingleCellExperiment}
\alias{get_coefs.SingleCellExperiment}
\title{Get simulated coefficient values for cluster and condition}
\usage{
\method{get_coefs}{SingleCellExperiment}(y)
}
\arguments{
\item{y}{A \code{SingleCellExperiment}}
}
\description{
Get simulated coefficient values for cluster and condition
}
