% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter.R
\name{get_conditions}
\alias{get_conditions}
\title{Get the treatment condition for each sample}
\usage{
get_conditions(sim)
}
\arguments{
\item{sim}{A \code{splatter} pop sim}
}
\description{
Get the treatment condition for each sample
}
