% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter.R
\name{get_groups.SingleCellExperiment}
\alias{get_groups.SingleCellExperiment}
\title{Get group labels from a \code{SingleCellExperiment}}
\usage{
\method{get_groups}{SingleCellExperiment}(y)
}
\arguments{
\item{y}{A \code{SingleCellExperiment}}
}
\description{
Get group labels from a \code{SingleCellExperiment}
}
