% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{get_pb_nf}
\alias{get_pb_nf}
\title{Get pseudobulk normalization factors}
\usage{
get_pb_nf(y)
}
\arguments{
\item{y}{A \code{DGEList} object of single cell counts with fitted dispersions}
}
\description{
Get pseudobulk normalization factors
}
