% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter.R
\name{get_pb_variables}
\alias{get_pb_variables}
\title{Get the design variables for pseudobulk samples}
\usage{
get_pb_variables(sim)
}
\arguments{
\item{sim}{A \code{splatter} pop sim}
}
\description{
Get the design variables for pseudobulk samples
}
