% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{create_pb_samples}
\alias{create_pb_samples}
\title{Aggregate single cell counts to create pseudobulk samples}
\usage{
create_pb_samples(y)
}
\arguments{
\item{y}{A \code{DGEList} object of single cell counts with pseudodata attached}
}
\description{
Aggregate single cell counts to create pseudobulk samples
}
